+++ src/griddyn/blocks/ControlBlock.cpp
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include "controlBlock.h"
+#include "ControlBlock.h"
 
 #include "core/coreObjectTemplates.hpp"
 #include "gmlc/utilities/vectorOps.hpp"
