+++ src/griddyn/blocks/DerivativeBlock.cpp
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include "derivativeBlock.h"
+#include "DerivativeBlock.h"
 
 #include "core/coreExceptions.h"
 #include "core/coreObjectTemplates.hpp"
