+++ src/griddyn/blocks/IntegralBlock.cpp
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include "integralBlock.h"
+#include "IntegralBlock.h"
 
 #include "core/coreObjectTemplates.hpp"
 #include "gmlc/utilities/vectorOps.hpp"
