+++ src/griddyn/blocks/blockLibrary.h
@@ -10,14 +10,14 @@
 @brief helper header file defining all the different block types
 */
 #include "blockSequence.h"
-#include "controlBlock.h"
-#include "deadbandBlock.h"
-#include "delayBlock.h"
-#include "derivativeBlock.h"
-#include "filteredDerivativeBlock.h"
-#include "functionBlock.h"
-#include "integralBlock.h"
-#include "lutBlock.h"
-#include "nullBlock.h"
-#include "pidBlock.h"
+#include "ControlBlock.h"
+#include "DeadbandBlock.h"
+#include "DelayBlock.h"
+#include "DerivativeBlock.h"
+#include "FilteredDerivativeBlock.h"
+#include "FunctionBlock.h"
+#include "IntegralBlock.h"
+#include "LutBlock.h"
+#include "NullBlock.h"
+#include "PidBlock.h"
 #include "transferFunctionBlock.h"
