+++ src/griddyn/primary/acBus.cpp
@@ -10,7 +10,7 @@
 #include "../GridArea.h"
 #include "../Link.h"
 #include "../Load.h"
-#include "../blocks/derivativeBlock.h"
+#include "../blocks/DerivativeBlock.h"
 #include "../simulation/contingency.h"
 #include "core/coreExceptions.h"
 #include "core/coreObjectTemplates.hpp"
